backend/requirements.txt
aniso8601>=9.0.1
Click>=8.0.0
Flask>=2.0.0
Flask-Cors>=3.0.10
Flask-RESTful>=0.3.9
Flask-SQLAlchemy>=2.5.1
itsdangerous>=2.0.0
Jinja2>=3.0.0
MarkupSafe>=2.0.0
psycopg2-binary>=2.9.0
pytz>=2021.1
six>=1.16.0
SQLAlchemy>=1.4.0
Werkzeug>=2.0.0

python-dotenv~=1.0.1